Remove default URLSessionHTTPClient initializer parameter (dependeny)
as clients should be mindful about which `URLSession` instance to use. For example, clients may decide to use the `.shared` instance, but when testing we can inject a `URLSession` with an ephemeral configuration registered with the `URLProtocolStub` class. This way, we don't need to start/stop intercepting URL requests globally.